Move curl method to utils/curl.rb.
