+++ Library/Homebrew/utils.rb
@@ -8,6 +8,7 @@ require "utils/fork"
 require "utils/git"
 require "utils/analytics"
 require "utils/github"
+require "utils/curl"
 
 class Tty
   class << self
@@ -341,21 +342,6 @@ def quiet_system(cmd, *args)
     $stdout.reopen("/dev/null")
     $stderr.reopen("/dev/null")
   end
-end
-
-def curl(*args)
-  curl = Pathname.new ENV["HOMEBREW_CURL"]
-  curl = Pathname.new "/usr/bin/curl" unless curl.exist?
-  raise "#{curl} is not executable" unless curl.exist? && curl.executable?
-
-  flags = HOMEBREW_CURL_ARGS
-  flags = flags.delete("#") if ARGV.verbose?
-
-  args = [flags, HOMEBREW_USER_AGENT_CURL, *args]
-  args << "--verbose" if ENV["HOMEBREW_CURL_VERBOSE"]
-  args << "--silent" if !$stdout.tty? || ENV["TRAVIS"]
-
-  safe_system curl, *args
 end
 
 def puts_columns(items)
